langgraph_app/requirements.txt
langgraph
langchain-openai
fastapi
uvicorn
azure-ai-formrecognizer
python-dotenv
pydantic
python-multipart